pyproject.toml
[project]
name = "mlx-image"
version = "0.1.10"
description = "Apple MLX image models library"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Riccardo Musmeci", email = "riccardomusmeci92@gmail.com" },
]
dependencies = [
    "mlx>=0.31.0",
    "numpy>=2.0",
    "tqdm>=4.66",
    "safetensors>=0.4",
    "opencv-python>=4.10",
    "Pillow>=12.1.1",
    "pandas>=2.2",
    "matplotlib>=3.9",
    "albumentations>=2.0",
    "huggingface_hub>=0.23",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3",
    "ruff>=0.11",
    "ty>=0.0.15",
    "detect-secrets>=1.5.0",
]
test = [
    "pytest>=9.0",
    "pytest-cov>=6.0",
    "pytest-sugar>=1.0",
    "pytest-xdist>=3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlxim"]

[tool.ruff]
src = ["src"]
line-length = 120
fix = true
target-version = "py313"

exclude = [
    ".bzr", ".direnv", ".eggs", ".git", ".hg", ".nox", ".pants.d",
    ".ruff_cache", ".svn", ".tox", ".venv", "__pypackages__", "_build",
    "buck-out", "build", "dist", "node_modules", "venv", "__pycache__",
    "bin", "lib", "local", "notebooks",
]

[tool.ruff.lint]
select = ["B", "C", "E", "F", "W", "I001", "I002", "UP"]
ignore = ["E501", "D103", "F401", "C901", "B905", "E741", "F403", "F405"]

[tool.ruff.lint.isort]
known-first-party = ["mlxim"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.coverage.run]
branch = true
dynamic_context = "test_function"
source = ["src/mlxim"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self\\.debug",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.coverage.html]
directory = "htmlcov"
show_contexts = true

[tool.local-ci]
skip_tests = true   # Requires Apple Silicon (MLX) - tests run manually on mini1/mini2
skip_ty = true      # External repo - not our type errors to fix

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::FutureWarning:mlxim.utils.validation",
]
log_file = "tests/logs/pytest-logs.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s | %(levelname)-8s | %(module)-20s | %(funcName)-20s | %(thread)-8d ; %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ty.environment]
python-version = "3.14"

[tool.ty.rules]
unresolved-import = "ignore"
possibly-missing-attribute = "warn"
